Finished TaskPool, SynchronizedSwimming, module complete

+++ src/_04_Thread_Pool/ThreadPoolTest.java
@@ -6,7 +6,7 @@ import java.util.Random;
 
 import org.junit.Test;
 
-import thread_pool.ThreadPool;
+import _04_Thread_Pool.ThreadPool;
 //   A Thread Pool is a way of managing multiple threads with multiple workloads.
 //   It works by collecting a list of tasks to be performed by a thread.
 //   A pool of threads is then created, and each thread takes a task from the list and does that work.
@@ -20,10 +20,10 @@ public class ThreadPoolTest {
 
 	@Test
 	public void test() {
-		int total = 1000000;
+		long total = 1000000;
 		
 		ThreadPool tp = new ThreadPool(4);
-		char[] chars = new char[total];
+		char[] chars = new char[(int) total];
 		
 		for(int i = 0; i < total; i++){
 			chars[i] = (char)(new Random().nextInt(25) + 97);
